+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentMethodMessagingElement/PaymentMethodMessagingElement.swift
@@ -12,7 +12,7 @@ import SwiftUI
 import UIKit
 
 /// An element that provides a view with information about how a purchase could be paid for using Buy Now, Pay Later payment methods.
-@_spi(STP)
+@_spi(PaymentMethodMessagingElementPreview)
 public class PaymentMethodMessagingElement {
 
     /// A UIKit view of the element.
@@ -107,13 +107,15 @@ public class PaymentMethodMessagingElement {
 
     let mode: Mode
     let infoUrl: URL
+    let legalDisclosure: String?
     let promotion: String
     let appearance: Appearance
     let analyticsHelper: PMMEAnalyticsHelper
 
-    init(mode: Mode, infoUrl: URL, promotion: String, appearance: PaymentMethodMessagingElement.Appearance, analyticsHelper: PMMEAnalyticsHelper) {
+    init(mode: Mode, infoUrl: URL, legalDisclosure: String?, promotion: String, appearance: PaymentMethodMessagingElement.Appearance, analyticsHelper: PMMEAnalyticsHelper) {
         self.mode = mode
         self.infoUrl = infoUrl
+        self.legalDisclosure = legalDisclosure
         self.promotion = promotion
         self.appearance = appearance
         self.analyticsHelper = analyticsHelper